news_podcast_generator/backend/requirements.txt
addict==2.4.0
agno==1.4.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioredis==2.0.1
aiosignal==1.3.2
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
APScheduler==3.11.0
asgi-csrf==0.11
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
billiard==4.2.1
blis==1.2.1
boto3==1.38.3
botocore==1.38.3
browser-use==0.2.4
cachetools==5.5.2
catalogue==2.0.10
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-default-group==1.2.4
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
csvw==3.5.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
Cython==3.0.12
datasette==0.65.1
defusedxml==0.7.1
distro==1.9.0
dlinfo==2.0.0
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
duckduckgo_search==8.0.1
tavily-python==0.3.0
crawl4ai==0.3.0
nest-asyncio==1.6.0
elevenlabs==1.57.0
espeakng-loader==0.2.4
faiss-cpu==1.9.0.post1
fastapi==0.115.12
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
flexcache==0.3
flexparser==0.4
flower==2.0.1
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
gnews==0.4.1
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
greenlet==3.2.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
humanize==4.12.3
hupper==1.12.1
hyperframe==6.1.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
janus==2.0.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kokoro==0.9.4
kombu==5.5.3
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langcodes==3.5.0
langsmith==0.3.31
language-tags==1.2.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
linkify-it-py==2.0.3
llvmlite==0.44.0
loguru==0.7.3
lxml==5.4.0
lxml_html_clean==0.4.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
mem0ai==0.1.93
mergedeep==1.3.4
misaki==0.9.4
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
murmurhash==1.0.12
networkx==3.4.2
newspaper4k==0.9.3.1
nltk==3.9.1
num2words==0.5.14
numba==0.61.2
numpy==1.26.4
ollama==0.4.7
openai==1.74.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
phonemizer-fork==3.3.2
pillow==11.2.1
pip-chill==1.0.3
playwright==1.52.0
pluggy==1.6.0
pooch==1.8.2
portalocker==2.10.1
posthog==3.25.0
preshed==3.0.9
primp==0.15.0
prometheus_client==0.22.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
pydub==0.25.1
pyee==13.0.0
PyJWT==2.9.0
pyparsing==3.2.3
pyperclip==1.9.0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.14.2
rdflib==7.1.4
redis==5.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
screeninfo==0.8.1
segments==2.3.0
sentence-transformers==4.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.8.5
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
textual==3.2.0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tldextract==5.3.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.2.2
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.0
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
widgetsnbextension==4.0.14
wrapt==1.17.2
yarl==1.20.0
zstandard==0.23.0